Mark SocketConnectionTest::testExceptioOpenSocket() skipped
This test is not consistent, it's passing in PHP (different versions)
and used to pass on HHVM older version, but it's not passing in newer
versions of HHVM.
The reason is that (as mentioned in PHP documentation) detecing
connection errors for UDP sockets is not guaranteed, because of the
connectionless nature of UDP.

At the moment I'll just skip this test, since the behavior is not that
much helpful for the whole library, but it's also misguies users of the
library.
The truth is that even PHP won't throw exceptions when socket
connections fail, but issues a WARNING, and when running tests with
PHPUnit we're treating WARNINGs as exceptions.

So maybe a better test is to test if this method issues a warning
instead of throwing exceptions.