Improve detecting if time delta is a callable in Telegraf timer command
Not all versions of PHP accept static methods the same. The better way
of specifying an static method is by passing an array of (class, method)
but php's builtin is_callable() had issues with this in version 5.3.
So instead the timer was accepting static methods as a string, but
by manually checking if the callable is an array it works OK for all
versions.

So now static methods should be passed as arrays to timing.