Telegraf client sets default tags for commands from its tags settings
And rename the tags settings to 'default_tags' to imply that this is
going to be the default tags, and they might be overriden by specific
metric calls.
Also provide a method to getDefaultTags() from the client,
so users can merge them with specific tags if they wanted to include the
default tags as well for a given metric.