Telegraf client merges default tags with metric tags by default
Previously the client used to use the default tags whenever a metric was
sent without a tag and ignore the default tags if the metric had tags.
So if the users needed to have the default tags included in the current
metric tags, they should merge the default tags and metric tags
manually, each time.

Now this commit changes the default behavior, so from now each metric
includes all the tags, including the default tags and current metric
tags.

This can be controlled by the new 'merge_tags' settings, passed to the
client constructor.