Telegraf client Timer command timing() triggers notice on callable param
The timing() method of the timer command used to accept a callable
instead of the duration, so the duration of the execution of the
callable was calculated internally and sent.
But this is not a good idea to change the behavior of the method based
on the type of the argument. Instead we're encouraging users to
call timeCallable() explicitly instead.